Refactor layout and sidebar components; - remove unused files and styles, and enhance variable management in SCSS. -Add 'USE' in scss files -Change Alt's of logos -Delete hardcoded colors -Add arial labels for FontAwesomeIcon
