+++ src/App.scss
@@ -1,5 +1,3 @@
-$primary-color: #ffd700;
-
 @import 'animate.css';
 
 @font-face {
