+++ src/components/Layout/Layout.tsx
@@ -1,5 +1,5 @@
-import Sidebar from '../Sidebar';
-import './style.scss';
+import Sidebar from '../Sidebar/Sidebar';
+import './Layout.module.scss';
 
 const Layout = () => {
   return <Sidebar></Sidebar>;
