+++ src/components/Sidebar/Sidebar.module.scss
@@ -1,10 +1,13 @@
-.nav-bar {
-  background-color: #18181852;
-  position: absolute;
-  top: 0;
+@use '../../styles/abstracts/variables';
+
+.NavBar {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
   width: 165px;
-  height: 100%;
+  height: 100vh;
   min-height: 500px;
+  background-color: #18181852;
   z-index: 3;
 
   .logo {
@@ -18,24 +21,21 @@
       width: 100%;
       height: auto;
 
-      &.sub-logo {
+      &.SubLogo {
         width: 165px;
       }
     }
   }
 
   nav {
-    position: absolute;
-    top: 50%;
     width: 100%;
-    height: 210px;
-    margin-top: -120px;
+    margin-top: 120px;
     display: block;
     text-align: center;
 
     a {
       font-size: 22px;
-      color: #4d4d4e;
+      color: variables.$gold-color;
       display: block;
       line-height: 51px;
       height: 51px;
@@ -43,8 +43,8 @@
       text-decoration: none;
 
       &:hover {
-        // color: $primary-color;
-        // color: #4d4d4e;
+        color: variables.$primary-color;
+        background-color: variables.$background-color-light;
       }
 
       i {
