+++ src/styles/abstracts/_variables.scss
@@ -1,0 +1,15 @@
+// COLORS
+$primary-color: #ffd700;
+$gold-color: #cc9b57;
+$secondary-color: #4d4d4e;
+$background-color: #18181852;
+$background-color-light: #142534;
+$background-color-dark: #0d0d0d;
+
+// FONT COLORS
+$fontColor: #ffffff;
+$font-icon-color: #4d4d4d;
+$font-color-light: #d9d9d9;
+$font-color-dark: #000000;
+$font-color-secondary: #4d4d4e;
+$font-color-secondary-light: #b3b3b3;
