+++ src/vite-env.d.ts
@@ -1,12 +1,2 @@
 /// <reference types="vite/client" />
 /// <reference types="vite-plugin-svgr/client" />
-
-declare module './reportWebVitals' {
-  const reportWebVitals: (onPerfEntry?: (entry: any) => void) => void;
-  export default reportWebVitals;
-}
-
-declare module './reportWebVitals' {
-  const reportWebVitals: (onPerfEntry?: (entry: any) => void) => void;
-  export default reportWebVitals;
-}
